+++ plugin-dev/Source/Sentry/Private/Apple/AppleSentrySubsystem.cpp
@@ -56,6 +56,11 @@ void FAppleSentrySubsystem::InitWithSettings(const USentrySettings* settings, co
 	USentryBeforeMetricHandler* beforeMetricHandler = callbackHandlers.BeforeMetricHandler;
 	USentryTraceSampler* traceSampler = callbackHandlers.TraceSampler;
 
+	if (callbackHandlers.BeforeSendFeedbackHandler != nullptr)
+	{
+		UE_LOG(LogSentrySdk, Warning, TEXT("beforeSendFeedback handler is set but isn't supported on Mac/iOS - it will be ignored."));
+	}
+
 	isScreenshotAttachmentEnabled = settings->AttachScreenshot;
 	isGameLogAttachmentEnabled = settings->EnableAutoLogAttachment;
 	isSessionReplayAttachmentEnabled = settings->AttachSessionReplay;