+++ plugin-dev/Source/Sentry/Private/GenericPlatform/GenericPlatformSentrySubsystem.cpp
@@ -17,6 +17,7 @@
 #include "SentryBeforeBreadcrumbHandler.h"
 #include "SentryBeforeLogHandler.h"
 #include "SentryBeforeMetricHandler.h"
+#include "SentryBeforeSendFeedbackHandler.h"
 #include "SentryBeforeSendHandler.h"
 #include "SentryBreadcrumb.h"
 #include "SentryDefines.h"
@@ -85,6 +86,18 @@ static void PrintVerboseLog(sentry_level_t level, const char* message, va_list a
 	if (closure)
 	{
 		return StaticCast<FGenericPlatformSentrySubsystem*>(closure)->OnBeforeSend(event, hint, closure, false);
+	}
+	else
+	{
+		return event;
+	}
+}
+
+/* static */ sentry_value_t FGenericPlatformSentrySubsystem::HandleBeforeSendFeedback(sentry_value_t event, sentry_hint_t* hint, void* closure)
+{
+	if (closure)
+	{
+		return StaticCast<FGenericPlatformSentrySubsystem*>(closure)->OnBeforeSendFeedback(event, hint, closure);
 	}
 	else
 	{
@@ -193,6 +206,44 @@ sentry_value_t FGenericPlatformSentrySubsystem::OnBeforeSend(sentry_value_t even
 	USentryEvent* ProcessedEvent = Handler->HandleBeforeSend(EventToProcess, nullptr);
 
 	return ProcessedEvent ? event : sentry_value_new_null();
+}
+
+sentry_value_t FGenericPlatformSentrySubsystem::OnBeforeSendFeedback(sentry_value_t event, sentry_hint_t* hint, void* closure)
+{
+	if (!closure || this != closure)
+	{
+		return event;
+	}
+
+	USentryBeforeSendFeedbackHandler* Handler = GetBeforeSendFeedbackHandler();
+	if (!Handler)
+	{
+		// If custom handler isn't set skip further processing
+		return event;
+	}
+
+	if (!SentryCallbackUtils::IsCallbackSafeToRun())
+	{
+		return event;
+	}
+
+	TSentryCallbackGuard<USentryBeforeSendFeedbackHandler> ReentrancyGuard;
+	if (ReentrancyGuard.IsReentrant())
+	{
+		return event;
+	}
+
+	USentryEvent* EventToProcess = USentryEvent::Create(MakeShareable(new FGenericPlatformSentryEvent(event, false)));
+
+	USentryEvent* ProcessedEvent = Handler->HandleBeforeSendFeedback(EventToProcess, nullptr);
+
+	if (!ProcessedEvent)
+	{
+		sentry_value_decref(event);
+		return sentry_value_new_null();
+	}
+
+	return event;
 }
 
 sentry_value_t FGenericPlatformSentrySubsystem::OnBeforeBreadcrumb(sentry_value_t breadcrumb, void* closure)
@@ -468,6 +519,7 @@ void FGenericPlatformSentrySubsystem::AttachStackTrace(sentry_value_t target)
 FGenericPlatformSentrySubsystem::FGenericPlatformSentrySubsystem()
 	: bUseNativeBackend(false)
 	, beforeSend(nullptr)
+	, beforeSendFeedback(nullptr)
 	, beforeBreadcrumb(nullptr)
 	, beforeLog(nullptr)
 	, beforeMetric(nullptr)
@@ -489,6 +541,7 @@ void FGenericPlatformSentrySubsystem::InitWithSettings(const USentrySettings* se
 	bUseNativeBackend = settings->UseNativeBackend;
 
 	beforeSend = callbackHandlers.BeforeSendHandler;
+	beforeSendFeedback = callbackHandlers.BeforeSendFeedbackHandler;
 	beforeBreadcrumb = callbackHandlers.BeforeBreadcrumbHandler;
 	beforeLog = callbackHandlers.BeforeLogHandler;
 	beforeMetric = callbackHandlers.BeforeMetricHandler;
@@ -581,6 +634,7 @@ void FGenericPlatformSentrySubsystem::InitWithSettings(const USentrySettings* se
 	sentry_options_set_sample_rate(options, settings->SampleRate);
 	sentry_options_set_max_breadcrumbs(options, settings->MaxBreadcrumbs);
 	sentry_options_set_before_send(options, HandleBeforeSend, this);
+	sentry_options_set_before_send_feedback(options, HandleBeforeSendFeedback, this);
 	sentry_options_set_before_send_log(options, HandleBeforeLog, this);
 	sentry_options_set_on_crash(options, HandleOnCrash, this);
 	sentry_options_set_shutdown_timeout(options, settings->ShutdownTimeout);
@@ -1179,6 +1233,11 @@ TSharedPtr<ISentryTransactionContext> FGenericPlatformSentrySubsystem::ContinueT
 USentryBeforeSendHandler* FGenericPlatformSentrySubsystem::GetBeforeSendHandler() const
 {
 	return beforeSend;
+}
+
+USentryBeforeSendFeedbackHandler* FGenericPlatformSentrySubsystem::GetBeforeSendFeedbackHandler() const
+{
+	return beforeSendFeedback;
 }
 
 USentryBeforeBreadcrumbHandler* FGenericPlatformSentrySubsystem::GetBeforeBreadcrumbHandler() const